podman

+++ src/cloud-init/podman_traefik.sh
@@ -3,97 +3,20 @@
 ## This is a wrapper script that creates another script at ${SCRIPT_INSTALL_PATH}.
 ##   ( /usr/local/sbin/podman_traefik.sh by default )
 ## The new installed script will have hard-coded the configuration gathered by
-## this wrapper script. If you need to update the config, edit the new script,
-## and re-run.
+## this wrapper script. If you need to update the config after installation,
+## edit the new script, and re-run.
 
-core_config() {
-    # Podman and Traefik config.
-    ## Permanent install path for the new script:
-    DEFAULT_SCRIPT_INSTALL_PATH=/usr/local/sbin/podman_traefik.sh
-
-    ## Traefik:
-    DEFAULT_TLS_ON=true
-    DEFAULT_DOMAIN=dev1.example.com
-    DEFAULT_ACME_EMAIL=you@example.com
-    DEFAULT_ACME_CA=https://acme-v02.api.letsencrypt.org/directory
-
-    ## Create array of all config variable names:
-    ## Other configs may also add to this list:
-    ALL_VARS=( SCRIPT_INSTALL_PATH \
-               TLS_ON \
-               DOMAIN \
-               ACME_EMAIL \
-               ACME_CA )
-}
-
-merge_config() {
-    ## load variables from the environment, or use the DEFAULT if unspecified:
-    # This is to be run after all all other configs have added vars to ALL_VARS
-    echo "## Config:"
-    for var in "${ALL_VARS[@]}"; do
-        default_name=DEFAULT_$var
-        default_value=${!default_name}
-        value=${!var:-$default_value}
-        declare -g $var=$value
-        echo $var=${!var}
-    done
-}
- 
-create_script() {
-    ## Save config in permanent script:
-    touch ${SCRIPT_INSTALL_PATH} && chmod 0700 ${SCRIPT_INSTALL_PATH}
-    ## Do the header first, which includes config:
-    cat <<'END_OF_SCRIPT_HEADER' > ${SCRIPT_INSTALL_PATH}
-#!/bin/bash -ex
-## Podman systemd config for Ubuntu 20.10
-
-## Default values that were used during first install:
-default_config() {
-END_OF_SCRIPT_HEADER
-
-    for var in "${ALL_VARS[@]}"; do
-        echo "    DEFAULT_$var=${!var}" >> ${SCRIPT_INSTALL_PATH}
-    done
-
-    cat <<'END_OF_SCRIPT_CONFIG' >> ${SCRIPT_INSTALL_PATH}
-}
-
-config() {
-    ## This config overrides the default config, and is loaded from the outside
-    ## environment variables. The names of the variables are the same as in the
-    ## default config, except without the `DEFAULT_` prefix. If no such variable
-    ## is provided, the default value is used instead.
-    export HOME=${HOME:-/root}
-END_OF_SCRIPT_CONFIG
-
-    for var in "${ALL_VARS[@]}"; do
-        echo "    $var=\${$var:-\$DEFAULT_$var}" >> ${SCRIPT_INSTALL_PATH}
-    done
-
-    cat <<'END_OF_INSTALLER' >> ${SCRIPT_INSTALL_PATH}
-}
-
-
-install() {
-    ## Create /etc/sysconfig to store container environment files
-    mkdir -p /etc/sysconfig
-
-    ## Install packages:
-    export DEBIAN_FRONTEND=noninteractive
-    apt-get update
-    apt-get -y install podman runc
-
-    ## Niceties:
-    echo "set enable-bracketed-paste on" >> /root/.inputrc
-}
-
-write_container_service() {
-    ## Template for a podman container systemd unit
+create_service_container() {
+    ## Template function to create a systemd unit for a podman container
     ## Expects environment file at /etc/sysconfig/${SERVICE}
     SERVICE=$1
     IMAGE=$2
     PODMAN_ARGS=$3
     CMD_ARGS=${@:4}
+    if [ ! -f /etc/sysconfig/${SERVICE} ]; then
+        echo "No environment file found: /etc/sysconfig/${SERVICE}"
+        exit 1
+    fi
     cat <<EOF > /etc/systemd/system/${SERVICE}.service
 [Unit]
 After=network-online.target
@@ -111,17 +34,66 @@ WantedBy=network-online.target
 EOF
 }
 
-traefik_service() {
-    SERVICE=traefik
-    IMAGE=traefik:v2.3
-    touch /etc/sysconfig/${SERVICE}
-    mkdir /etc/sysconfig/${SERVICE}.d
-    write_container_service \
-      ${SERVICE} ${IMAGE} "-v /etc/sysconfig/${SERVICE}.d:/etc/traefik/" \
-      --providers.file.directory=/etc/traefik \
-      --providers.file.watch=true
+create_service_proxy() {
+    ## Template function to create a traefik configuration for a service.
+    ## Creates automatic http to https redirect.
+    ## Note: traefik will automatically reload configuration when the file changes.
+    TRAEFIK_SERVICE=traefik
+    SERVICE=$1
+    DOMAIN=$2
+    cat <<END_TRAEFIK_CONF > /etc/sysconfig/${TRAEFIK_SERVICE}.d/${SERVICE}.toml
+[http.routers.${SERVICE}]
+  entrypoints = "web"
+  rule = "Host(\"${DOMAIN}\")"
+  middlewares = "${SERVICE}-secure-redirect"
+[http.middlewares.${SERVICE}-secure-redirect.redirectscheme]
+  scheme = "https"
+  permanent = "true"
+[http.routers.${SERVICE}-secure]
+  entrypoints = "websecure"
+  rule = "Host(\"${DOMAIN}\")"
+  service = "${SERVICE}"
+  tls = "true"
+  [http.routers.${SERVICE}-secure.tls]
+    certresolver = "default"
+END_TRAEFIK_CONF
+}
 
-    cat <<END_TRAEFIK_CONF > /etc/sysconfig/${SERVICE}.d/traefik.toml
+
+main() {
+    core_config() {
+        # Podman and Traefik config.
+        ## Permanent install path for the new script:
+        DEFAULT_SCRIPT_INSTALL_PATH=/usr/local/sbin/podman_traefik.sh
+
+        ## Traefik:
+        DEFAULT_TRAEFIK_IMAGE=traefik:v2.3
+        DEFAULT_DOMAIN=dev1.example.com
+        DEFAULT_ACME_EMAIL=you@example.com
+        DEFAULT_ACME_CA=https://acme-v02.api.letsencrypt.org/directory
+
+        ## Required output variables:
+        ##  - Create array of all TEMPLATES (functions) for this config:
+        ##  - Create array of all config VARS from this config:
+        TEMPLATES=(traefik_service)
+        VARS=( SCRIPT_INSTALL_PATH \
+               TRAEFIK_IMAGE \
+               DOMAIN \
+               ACME_EMAIL \
+               ACME_CA )
+    }
+
+    traefik_service() {
+        SERVICE=traefik
+        IMAGE=${TRAEFIK_IMAGE}
+        touch /etc/sysconfig/${SERVICE}
+        mkdir /etc/sysconfig/${SERVICE}.d
+        create_service_container \
+            ${SERVICE} ${IMAGE} "-v /etc/sysconfig/${SERVICE}.d:/etc/traefik/" \
+            --providers.file.directory=/etc/traefik \
+            --providers.file.watch=true
+
+        cat <<END_TRAEFIK_CONF > /etc/sysconfig/${SERVICE}.d/traefik.toml
 [log]
   level = "DEBUG"
 [api]
@@ -142,7 +114,79 @@ traefik_service() {
   email=${ACME_EMAIL}
 END_TRAEFIK_CONF
 
-    systemctl enable --now ${SERVICE}
+        systemctl enable --now ${SERVICE}
+    }
+
+
+    merge_config() {
+        ## load variables from the environment, or use the DEFAULT if unspecified:
+        # This is to be run after all all other configs have added vars to ALL_VARS
+        echo "## Config:"
+        for var in "${ALL_VARS[@]}"; do
+            default_name=DEFAULT_$var
+            default_value=${!default_name}
+            value=${!var:-$default_value}
+            declare -g $var=$value
+            echo $var=${!var}
+        done
+    }
+
+    create_script() {
+        ## Save config in permanent script:
+        touch ${SCRIPT_INSTALL_PATH} && chmod 0700 ${SCRIPT_INSTALL_PATH}
+        ## Do the header first, which includes hard-coded config:
+        cat <<'END_OF_SCRIPT_HEADER' > ${SCRIPT_INSTALL_PATH}
+#!/bin/bash -ex
+## Podman systemd config for Ubuntu 20.10
+
+## Default values that were used during first install:
+default_config() {
+END_OF_SCRIPT_HEADER
+
+        for var in "${ALL_VARS[@]}"; do
+            echo "    DEFAULT_$var=${!var}" >> ${SCRIPT_INSTALL_PATH}
+        done
+
+        cat <<'END_OF_SCRIPT_CONFIG' >> ${SCRIPT_INSTALL_PATH}
+}
+
+config() {
+    ## This config overrides the default config, and is loaded from the outside
+    ## environment variables. The names of the variables are the same as in the
+    ## default config, except without the `DEFAULT_` prefix. If no such variable
+    ## is provided, the default value is used instead.
+    export HOME=${HOME:-/root}
+END_OF_SCRIPT_CONFIG
+
+        for var in "${ALL_VARS[@]}"; do
+            echo "    $var=\${$var:-\$DEFAULT_$var}" >> ${SCRIPT_INSTALL_PATH}
+        done
+
+        cat <<END_DYNAMIC_CONFIG_1 >> ${SCRIPT_INSTALL_PATH}
+}
+
+$(declare -f create_service_container)
+$(declare -f create_service_proxy)
+END_DYNAMIC_CONFIG_1
+
+        for template in "${ALL_TEMPLATES[@]}"; do
+            cat <<END_DYNAMIC_CONFIG_2 >> ${SCRIPT_INSTALL_PATH}
+$(declare -f ${template})
+END_DYNAMIC_CONFIG_2
+        done
+
+        cat <<'END_OF_INSTALLER' >> ${SCRIPT_INSTALL_PATH}
+install_packages() {
+    ## Create /etc/sysconfig to store container environment files
+    mkdir -p /etc/sysconfig
+
+    ## Install packages:
+    export DEBIAN_FRONTEND=noninteractive
+    apt-get update
+    apt-get -y install podman runc
+
+    ## Niceties:
+    echo "set enable-bracketed-paste on" >> /root/.inputrc
 }
 
 main() {
@@ -152,10 +196,10 @@ main() {
     fi
     default_config
     config
-    install
-    traefik_service
-    postgres_service
-    phoenix_service
+    install_packages
+    for template in "${ALL_TEMPLATES[@]}"; do
+      $template
+    done
     chmod go-rwx -R /etc/sysconfig
     echo "All done :)"
 }
@@ -163,15 +207,23 @@ main() {
 main
 END_OF_INSTALLER
 
-    echo "## Script written to ${SCRIPT_INSTALL_PATH}"
-}
+        echo "## Script written to ${SCRIPT_INSTALL_PATH}"
+    }
 
-main() {
-    # Run core config first:
-    core_config
-    # Run each application config function next:
+    # Initialize list of all template functions
+    ALL_TEMPLATES=()
+    # Initialize list of all config variables
+    ALL_VARS=()
+    # Run all configs, core_config first:
+    ALL_CONFIGS=(core_config ${ALL_CONFIGS[@]})
     for var in "${ALL_CONFIGS[@]}"; do
+        TEMPLATES=()
+        VARS=()
+        ## Run the config (which sets TEMPLATES and VARS):
         $var
+        ## Append templates and vars:
+        ALL_SERVICES=(${ALL_TEMPLATES[@]} ${TEMPLATES[@]})
+        ALL_VARS=(${ALL_VARS[@]} ${VARS[@]})
     done
     # Merge all the configs, applying environment vars first, then the defaults:
     merge_config
